--- keyboards/doro67/readme.md
@@ -1,13 +1,0 @@
-# doro67
-
-65% keyboard with integrated plate, USB C, and thick bezels!
-
-Keyboard Maintainer: [MechMerlin](https://github.com/mechmerlin)  
-Hardware Supported: Doro67 
-Hardware Availability: [Geekhack GB](https://geekhack.org/index.php?topic=97265.0)
-
-Make example for this keyboard (after setting up your build environment):
-
-    make doro67:default
-
-See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
