+++ keyboards/doro67/regular/info.json
