+++ keyboards/doro67/regular/readme.md
@@ -1,0 +1,15 @@
+# Regular PCB for the Doro67 
+
+Regular PCB for the Doro67 having ONLY one layout. This is not the PCB with RGB support. 
+
+Do not flash RGB firmware for this board. 
+
+Keyboard Maintainer: [MechMerlin](https://github.com/mechmerlin)  
+Hardware Supported: Doro67 
+Hardware Availability: [Geekhack GB](https://geekhack.org/index.php?topic=97265.0)
+
+Make example for this keyboard (after setting up your build environment):
+
+    make doro67/regular:default
+
+See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
